pyproject.toml
[project]
name = "skillshub"
version = "0.1.0"
description = "Centralized skills directory for AI agent teams. Sync, update, and create agentskills.io skills across all your agents."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "gitpython>=3.1",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "mcp>=1.0",
]

[project.scripts]
skillshub = "skillshub.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skillshub"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
